// src/core/enums/type-of-dependence.enums.ts
export enum TypeOfDependenceEnum {
  Machinery = 1,
  AreaOfBuildingsAndFacilities = 2,
  Trees = 3,
  TotalArea = 4,
  Equipment= 5,

}

export enum TypeOfDependencePersianEnum {
  Machinery = "ماشین آلات و ادوات",
  AreaOfBuildingsAndFacilities = "ساختمان و تأسیسات",
  Trees = "درختان",
  TotalArea = "مساحت کل",
  Equipment = "تجهیزات",
}


// public enum TypeOfDependence
//     {
//         [Description("ماشین آلات")]
//         Machinery = 1,

//         [Description("مساحت ساختمان و تاسیسات")]
//         AreaOfBuildingsAndFacilities = 2,

//         [Description("درختان")]
//         Trees = 3,

//         [Description("مساحت کل")]
//         TotalArea = 4,
//     }